Move MQE codes to `server-core`.
