+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/mqe/rt/exception/IllegalExpressionException.java
@@ -16,29 +16,10 @@
  *
  */
 
-package org.apache.skywalking.oap.query.graphql.mqe.rt.operation.aggregatelabels;
+package org.apache.skywalking.oap.server.core.mqe.rt.exception;
 
-public class MinAggregateLabelsFunc implements AggregateLabelsFunc {
-
-    private Double min;
-
-    @Override
-    public void combine(final Double value) {
-        if (value == null) {
-            return;
-        }
-
-        if (min == null) {
-            min = value;
-        } else {
-            if (value < min) {
-                min = value;
-            }
-        }
-    }
-
-    @Override
-    public Double getResult() {
-        return min;
+public class IllegalExpressionException extends Exception {
+    public IllegalExpressionException(String message) {
+        super(message);
     }
 }
