+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/mqe/rt/MQEVisitor.java
@@ -25,17 +25,16 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.skywalking.oap.query.graphql.mqe.rt.exception.IllegalExpressionException;
-import org.apache.skywalking.oap.query.graphql.mqe.rt.operation.AggregateLabelsOp;
-import org.apache.skywalking.oap.query.graphql.mqe.rt.operation.AggregationOp;
 import org.apache.skywalking.oap.query.graphql.resolver.MetricsQuery;
 import org.apache.skywalking.oap.query.graphql.resolver.RecordsQuery;
-import org.apache.skywalking.oap.query.graphql.type.mql.ExpressionResult;
-import org.apache.skywalking.oap.query.graphql.type.mql.ExpressionResultType;
-import org.apache.skywalking.oap.query.graphql.type.mql.MQEValue;
-import org.apache.skywalking.oap.query.graphql.type.mql.MQEValues;
-import org.apache.skywalking.oap.query.graphql.type.mql.Metadata;
 import org.apache.skywalking.oap.server.core.Const;
+import org.apache.skywalking.oap.server.core.mqe.rt.MQEVisitorBase;
+import org.apache.skywalking.oap.server.core.mqe.rt.exception.IllegalExpressionException;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.ExpressionResult;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.ExpressionResultType;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.MQEValue;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.MQEValues;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.Metadata;
 import org.apache.skywalking.oap.server.core.query.DurationUtils;
 import org.apache.skywalking.oap.server.core.query.PointOfTime;
 import org.apache.skywalking.oap.server.core.query.enumeration.Order;
@@ -53,14 +52,9 @@ import org.apache.skywalking.oap.server.core.storage.annotation.Column;
 import org.apache.skywalking.oap.server.core.storage.annotation.ValueColumnMetadata;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.mqe.rt.grammar.MQEParser;
-import org.apache.skywalking.mqe.rt.grammar.MQEParserBaseVisitor;
-
-import static org.apache.skywalking.oap.query.graphql.mqe.rt.operation.BinaryOp.doBinaryOp;
-import static org.apache.skywalking.oap.query.graphql.mqe.rt.operation.FunctionOp.doFunction0Op;
-import static org.apache.skywalking.oap.query.graphql.mqe.rt.operation.FunctionOp.doFunction1Op;
 
 @Slf4j
-public class MQEVisitor extends MQEParserBaseVisitor<ExpressionResult> {
+public class MQEVisitor extends MQEVisitorBase {
     private final MetricsQuery metricsQuery;
     private final RecordsQuery recordsQuery;
     private final Entity entity;
@@ -76,171 +70,6 @@ public class MQEVisitor extends MQEParserBaseVisitor<ExpressionResult> {
         this.recordsQuery = recordsQuery;
         this.entity = entity;
         this.duration = duration;
-    }
-
-    @Override
-    public ExpressionResult visitAddSubOp(MQEParser.AddSubOpContext ctx) {
-        ExpressionResult left = visit(ctx.expression(0));
-        if (StringUtil.isNotBlank(left.getError())) {
-            return left;
-        }
-        ExpressionResult right = visit(ctx.expression(1));
-        if (StringUtil.isNotBlank(right.getError())) {
-            return right;
-        }
-        int opType = ctx.addSub().getStart().getType();
-        try {
-            return doBinaryOp(left, right, opType);
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitMulDivModOp(MQEParser.MulDivModOpContext ctx) {
-        ExpressionResult left = visit(ctx.expression(0));
-        if (StringUtil.isNotBlank(left.getError())) {
-            return left;
-        }
-        ExpressionResult right = visit(ctx.expression(1));
-        if (StringUtil.isNotBlank(right.getError())) {
-            return right;
-        }
-        int opType = ctx.mulDivMod().getStart().getType();
-        try {
-            return doBinaryOp(left, right, opType);
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitScalar(MQEParser.ScalarContext ctx) {
-        ExpressionResult result = new ExpressionResult();
-        double value = Double.parseDouble(ctx.getText());
-        MQEValue mqeValue = new MQEValue();
-        mqeValue.setDoubleValue(value);
-        mqeValue.setEmptyValue(false);
-        MQEValues mqeValues = new MQEValues();
-        mqeValues.getValues().add(mqeValue);
-        result.getResults().add(mqeValues);
-        result.setType(ExpressionResultType.SINGLE_VALUE);
-        return result;
-    }
-
-    @Override
-    public ExpressionResult visitAggregationOp(MQEParser.AggregationOpContext ctx) {
-        int opType = ctx.aggregation().getStart().getType();
-        ExpressionResult expResult = visit(ctx.expression());
-        if (StringUtil.isNotEmpty(expResult.getError())) {
-            return expResult;
-        }
-        try {
-            return AggregationOp.doAggregationOp(expResult, opType);
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitAggregateLabelsOp(final MQEParser.AggregateLabelsOpContext ctx) {
-        int funcType = ctx.aggregateLabelsFunc().getStart().getType();
-        ExpressionResult expResult = visit(ctx.expression());
-        if (StringUtil.isNotEmpty(expResult.getError())) {
-            return expResult;
-        }
-
-        if (!expResult.isLabeledResult()) {
-            expResult.setError("The result of expression [" + ctx.expression().getText() + "] is not a labeled result.");
-            return expResult;
-        }
-
-        try {
-            return AggregateLabelsOp.doAggregateLabelsOp(expResult, funcType);
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitFunction0OP(MQEParser.Function0OPContext ctx) {
-        int opType = ctx.function0().getStart().getType();
-        ExpressionResult expResult = visit(ctx.expression());
-        if (StringUtil.isNotEmpty(expResult.getError())) {
-            return expResult;
-        }
-        try {
-            return doFunction0Op(expResult, opType);
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitFunction1OP(MQEParser.Function1OPContext ctx) {
-        int opType = ctx.function1().getStart().getType();
-        ExpressionResult expResult = visit(ctx.expression());
-        if (StringUtil.isNotEmpty(expResult.getError())) {
-            return expResult;
-        }
-        try {
-            return doFunction1Op(expResult, opType, Integer.parseInt(ctx.parameter().INTEGER().getText()));
-        } catch (IllegalExpressionException e) {
-            ExpressionResult result = new ExpressionResult();
-            result.setType(ExpressionResultType.UNKNOWN);
-            result.setError(e.getMessage());
-            return result;
-        }
-    }
-
-    @Override
-    public ExpressionResult visitTopNOP(MQEParser.TopNOPContext ctx) {
-        return visit(ctx.metric());
-    }
-
-    @Override
-    public ExpressionResult visitRelablesOP(MQEParser.RelablesOPContext ctx) {
-        ExpressionResult result = visit(ctx.expression());
-        if (!result.isLabeledResult()) {
-            // Reserve the original result type
-            result.setError("The result of expression [" + ctx.expression().getText() + "] is not a labeled result.");
-            return result;
-        }
-
-        List<String> relabelList = Collections.emptyList();
-        String newLabelValue = ctx.label().labelValue().getText();
-        String labelValueTrim = newLabelValue.substring(1, newLabelValue.length() - 1);
-        if (StringUtil.isNotBlank(labelValueTrim)) {
-            relabelList = Arrays.asList(labelValueTrim.split(Const.COMMA));
-        }
-        List<MQEValues> mqeValuesList = result.getResults();
-
-        if (mqeValuesList.size() != relabelList.size()) {
-            // Resever the original result type
-            result.setError("The number of relabels is not equal to the number of labels.");
-            return result;
-        }
-        // For now, we only have a single label named `label`
-        for (int i = 0; i < mqeValuesList.size(); i++) {
-            mqeValuesList.get(i).getMetric().getLabels().get(0).setValue(relabelList.get(i));
-        }
-
-        return result;
     }
 
     @Override
