+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/resolver/MetricsExpressionQuery.java
@@ -25,9 +25,9 @@ import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.apache.skywalking.oap.query.graphql.mqe.rt.MQEVisitor;
-import org.apache.skywalking.oap.query.graphql.mqe.rt.exception.ParseErrorListener;
-import org.apache.skywalking.oap.query.graphql.type.mql.ExpressionResult;
-import org.apache.skywalking.oap.query.graphql.type.mql.ExpressionResultType;
+import org.apache.skywalking.oap.server.core.mqe.rt.exception.ParseErrorListener;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.ExpressionResult;
+import org.apache.skywalking.oap.server.core.mqe.rt.type.ExpressionResultType;
 import org.apache.skywalking.oap.server.core.query.input.Entity;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
